Fixed merchandises for Receive Line

+++ src/main/java/com/walmart/finance/ap/fds/receiving/response/ReceiveMDSResponse.java
@@ -12,6 +12,6 @@ import lombok.Setter;
 public class ReceiveMDSResponse {
     private Integer mdseDisplayCode;
     private Integer mdseConditionCode;
-    private Integer mdseQuantity;
+    private Long mdseQuantity;
     private String mdseQuantityUnitOfMeasureCode;
 }
